Feat: create the bg service outline for the feat
I created the outline of bg service for text-based monitoring feature
by putting all things enough to run the service once registered in the
Startup.cs.

So, the feat has bg service but is unable to run the remote script to
get the apps on a remote desktop. Thus also, DB related codes aren't
created yet to store monitored apps.